So far this code seems pretty solid. All the states are circled properly, deleted Serial.write calls that were used for testing.
Problem is the printer does not seem to be actually printing. It apparently
does the job, you can see the paper feeding, goes through all the stages,
but does not burn the paper. Data seems to be in the buffer just fine
so it's probably a subtle issue.

Next step is adding STATUS Serial.write after every package send.
I can probably turn them on with a debug flag, or maybe just leave them on
since it's always useful to know the specific status at which the GBPrinter
is (internally, whether they are errors, etc.)