Rolled back a few changed I had done in the previous commit to test stuff live. Applied a fix on the print command (margins had to be send on the second byte of the print data, mixed together. First payload byte fixed to 0x01)
Now that this is sorted out, printing actually works! And it seems that the state machine
is mostly good. It can even print multiple images in a row, without stalling or failures
of any kind.

Margins still not as expected, but I'll work on them. So far, just 0 margins
and then press the physical buttons on the gbprinter to roll the paper

Quite excited, to finally have a functional firmware!

Soon enough I'll post a client and I'll start working on an image converter, so that
any random image can actually be sent to the printer. Will be awesome.