This is a first commit of the client. I've been playing with this go cli client for the GBPrinter for a while, and so far, all the image transformation and serialization works correctly.
The cli client allows to take any jpg, gif or png image, rescale it,
convert to a 2-bit grayscale (color depth accepted by the gbprinter) and
either output the resulting file to a jpg (so you can check the results
of the transformations without having to print anything) or serialize the
output to a byte array that the gbprinter can properly print.

The serial interface needs work still. Once I get a working version of that
though, the client will be fully functional, and the project feature complete.